pyproject.toml
[tool.poetry]
name = "CDS-attendance"
version = "0.1.0"
description = "A modern agricultural trading platform API"
authors = ["Abdulalim <aaladan453@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlmodel = "^0.0.14"
pydantic = { extras = ["email"], version = "^2.6.0" }
pydantic-settings = "^2.1.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.6"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
email-validator = "^2.1.0"
authlib = "^1.3.0"
httpx = "^0.26.0"
pyyaml = "^6.0.1"
redis = "^6.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
locust = "^2.24.0"

[tool.poetry.scripts]
start = "app.main:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
